Prepare backend for deployment with updated controllers and repositories

+++ src/main/kotlin/factories/CustomerFactory.kt
@@ -6,8 +6,8 @@ import domain.entities.Position
 object CustomerFactory {
     fun fromCsvRow(row: List<String>): Customer {
         val id = row[0].toInt()
-        val x = row[1].toInt()
-        val y = row[2].toInt()
+        val x = row[1].toDouble()
+        val y = row[2].toDouble()
 
         val position = Position(x, y)
         return Customer(id, position)
